fix: circular dependency - transactions와 transactions_detail 간 순환 참조를 해결하기 위해 모델을 통합하였습니다.
